Yo,
Have been trying all day to figure out this change:

I want to be able to time stretch the loaded pieces... so the teacher can speed up or slow down their selected pieces by a factor of two. I should be able to do this with the audio manipulation library librosa.  For now, I want the BPM slider on the Plie page to reflect the values of the piece loaded into mem. After a piece is loaded, I want the slider value to change to the bpm of the piece, the minimum slider value to equal the bpm / 2, and the max to equal bpm * 2. I'm trying to implement the "update_bpm_slider" method from tap_tempo.py at the end of the "update_selected_piece" method in piece_selection.py but I can't crack it. Right now, it's giving me an attribute error: type object 'TapTempo' has no attribute 'bpm_slider' ... but it does have a bpm_slider attribute! Haha, very confused here.